TRANSLATOR WARNING! CHAPTER 5 CHANGES I'm sorry about this one guys. I had to change chapter 5 a fair bit and make chapter 6 also a little different.
I needed to highlight the new chapter5 folder in the supplied source code, which contains an existing sample application with behat tests run. This was pretty important to the book and people were having a hard time missing it. So, it'll be a little bit of a rewrite which I am sorry about, but it had to be done.