[chapter 12] Fix return HTTP status code
This fixes an example of content negotiation. When asked for an unsupported content type, the code answers with a 415 status code (Unsupported media type). However, a 406 (Not acceptable) would be more appropriate.

The descriptions of these codes say:

406 Not acceptable
The resource is valid, but cannot be provided in a format specified in the Accept headers in the request.

415 Unsupported media type
The client provided data with a media type that the server does not support.

In this example, the client is requesting data by specifying a content type in the `Accept` header, not providing data to the server.