Create eight-of-the-biggest-lies-in-APIs (#583)
* Create eight-of-the-biggest-lies-in-APIs

Submitting an article for the APIs You Won't Hate blog. Will need a cover image. In addition, @philsturgeon thought it might be worthwhile to springboard from this piece and link to other, existing articles (for example, could see the cache lie easily linking to the published work on caching). Somebody with knowledge of the deep cuts will need to add those, if desired.

* Update src/content/blog/eight-of-the-biggest-lies-in-APIs

Co-authored-by: Phil Sturgeon <67381+philsturgeon@users.noreply.github.com>

* add mdx extension

* look up slugs by lowercase file name

* add cover image

* Adjust FP cover size

* space out home page stack

* all lowercase file name swap pt.1

* all lower case file name swap pt.2

Co-authored-by: Mike Bifulco <mbifulco@live.com>
Co-authored-by: Phil Sturgeon <67381+philsturgeon@users.noreply.github.com>